requirements.txt
pandas>=2.0,<3.0
pyyaml>=6.0
pandera>=0.18
streamlit>=1.30
